+++ CHANGES.md
@@ -1,6 +1,7 @@
 #### unreleased changes
 
 * Move inline Javascript to standalone js files to reduce complexity.
+* Rename plugin to "JDITC Add Wordfence 2FA to Ultimate Member" and slug to "jditc-add-wordfence-2fa-to-ultimate-member" for WordPress.org compliance.
 
 #### 1.0 / 2026-07-07